+++ application/views/main/form.php
@@ -20,9 +20,9 @@
           <fieldset class="item pure-group pure-u-1">
             <input type="text" name="headline" id="item-headline" class="pure-input-1" placeholder="Headline*" value="<?php echo isset($_GET['headline'])?$_GET['headline']:(isset($item)?stripslashes($item->headline):''); ?>" required title="Max length 255 characters.">
             <?php if($type == 'headline'): ?>
-              <textarea name="notes" id="item-notes" class="pure-input-1" placeholder="Description"><?php echo isset($item->notes)?stripcslashes($item->notes):''; ?></textarea>
+              <textarea name="notes" id="item-notes" class="pure-input-1" placeholder="Description"><?php echo isset($item->notes)?stripcslashes(str_replace('\r', '', str_replace('\n', '', $item->notes))):''; ?></textarea>
             <?php elseif($type == 'article'): ?>
-              <textarea name="article" id="item-article" class="pure-input-1" placeholder="Article"><?php echo isset($item->article)?stripslashes($item->article):''; ?></textarea>
+              <textarea name="article" id="item-article" class="pure-input-1" placeholder="Article"><?php echo isset($item->article)?stripslashes(str_replace('\r', '', str_replace('\n', '', $item->article))):''; ?></textarea>
             <?php endif; ?>
           </fieldset>
         </div>
@@ -40,7 +40,7 @@
           <fieldset class="images pure-group pure-u-1">
             <?php if(isset($images_output) && $images_output): ?>
               <?php foreach($images_output as $i): ?>
-                <div class="pure-g">
+                <div class="pure-g <?php echo (@getimagesize($i->image))?'':'hidden'; ?>">
                   <div class="pure-u-1-24">
                     <div class="horizontal-margin-tiny vertical-margin-xsmall">
                       <img src="<?php echo stripslashes($i->image); ?>">