+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 <h1>Changelog</h1>
 
+<h2>Version 362</h2>
+
+<h3>Fix:</h3>
+* Mention Comment Notification issue
+* issue with \n on article/note output
+
 <h2>Version 361</h2>
 
 <h3>Fix:</h3>